port to 64 bit windows

+++ libraries/ObjectIO/ObjectIO/OS Windows/Windows_C_12/cCrossCallClipboard_121.c
@@ -77,7 +77,7 @@ void EvalCcRqGETCLIPBOARDTEXT (CrossCallInfo *pcci)			/* no params; string resul
 	}
 	CloseClipboard ();
 
-	MakeReturn1Cci (pcci, (int) result);
+	MakeReturn1Cci (pcci, (size_t) result);
 }
 
 /*	CcRqGETCLIPBOARDCOUNT is used by Clean to retrieve current clipboard count.
